pull search input and image creation out of chatSection
into its own module, user-management
