+++ src/js/elements/elements.js
@@ -55,6 +55,14 @@ class Element {
         return document.createElement(this.tagType.INPUT);
     }
 
+    // todo test this
+    createTextInput() {
+        let chatText = this.createDiv();
+        this._setClass(chatText, [className.TEXT]);
+        chatText.setAttribute('contenteditable', true);
+        return chatText;
+    }
+
     createUl() {
         return document.createElement(this.tagType.UL);
     }
